+++ src/devops/config/__init__.py
@@ -1,6 +1,7 @@
 """DevOps config package."""
 
 from .config import init_config
+from .config_git import GitConfig
 from .constants import Constants
 
-__all__ = ["Constants", "init_config"]
+__all__ = ["Constants", "GitConfig", "init_config"]